+++ pkg/scheduler/service/bookkeeper.go
@@ -110,62 +110,7 @@ func (bk *bookkeeper) Run(ctx context.Context, tasks ...BookkeepingTask) error {
 	}
 }
 
-type BookkeepingTask interface {
-	Apply(*ReconciliationResult) error
-}
 
-type orphanOperation struct{
-	transition *ClusterStatusTransition
-	logger *zap.SugaredLogger
-}
-
-func (oo *orphanOperation) Apply(reconResult *ReconciliationResult) error {
-	for _, orphanOp := range reconResult.GetOrphans() {
-		if orphanOp.State == model.OperationStateOrphan {
-			//don't update orphan operations which are already marked as 'orphan'
-			continue
-		}
-
-		if err := oo.transition.reconRepo.UpdateOperationState(
-			orphanOp.SchedulingID, orphanOp.CorrelationID, model.OperationStateOrphan); err == nil {
-			oo.logger.Infof("orphanOperation marked operation '%s' as orphan: "+
-				"last update %.2f minutes ago)", orphanOp, time.Since(orphanOp.Updated).Minutes())
-		} else {
-			oo.logger.Errorf("Bookkeeper failed to update status of orphan operation '%s': %s",
-				orphanOp, err)
-		}
-	}
-	return nil // TODO: Error richtig returnen, bis jtezt wirds nur nil returned, schauen wie man das am besten umsetzen kann
-}
-
-type finishOperation struct {
-	transition *ClusterStatusTransition
-	logger *zap.SugaredLogger
-}
-
-func (fo *finishOperation) Apply(reconResult *ReconciliationResult) error {
-	recon := reconResult.Reconciliation()
-	newClusterStatus := reconResult.GetResult()
-	errMsg := fmt.Sprintf("finishOperation failed to update cluster '%s' to status '%s' "+
-		"(triggered by reconciliation with schedulingID '%s'): CLuster is already in final state",
-		recon.RuntimeID, newClusterStatus, recon.SchedulingID)
-
-	if newClusterStatus.IsFinal() {
-		err := fo.transition.FinishReconciliation(recon.SchedulingID, newClusterStatus)
-		if err == nil {
-			fo.logger.Infof("finishOperation updated cluster '%s' to status '%s' "+
-				"(triggered by reconciliation with schedulingID '%s')",
-				recon.RuntimeID, newClusterStatus, recon.SchedulingID)
-			return nil
-		}
-		errMsg = fmt.Sprintf("finishOperation failed to update cluster '%s' to status '%s' "+
-			"(triggered by reconciliation with schedulingID '%s'): %s",
-			recon.RuntimeID, newClusterStatus, recon.SchedulingID, err)
-		fo.logger.Errorf(errMsg)
-	}
-
-	return errors.New(errMsg)
-}
 
 func (bk *bookkeeper) newReconciliationResult(recon *model.ReconciliationEntity) (*ReconciliationResult, error) {
 	ops, err := bk.repo.GetOperations(recon.SchedulingID)
